Array Rotation and OddOcurrencesInArray
